Refactoring authentication code for UTs
