+++ azure-devops/azext_devops/test/boards/test_workitem.py
@@ -16,22 +16,21 @@ from azext_devops.dev.boards.work_item import (delete_work_item,
                                             show_work_item)
 from azext_devops.dev.common.services import clear_connection_cache
 from azext_devops.test.utils.helper import get_client_mock_helper, TEST_DEVOPS_ORG_URL
+from azext_devops.test.utils.authentication import AuthenticatedTests
 
 
 
-class TestWorkItemMethods(unittest.TestCase):
+class TestWorkItemMethods(AuthenticatedTests):
 
     _TEST_DEVOPS_ORGANIZATION = TEST_DEVOPS_ORG_URL
     _TEST_PAT_TOKEN = 'somepat'
 
     def setUp(self):
-
+        self.authentication_setUp()
         self.get_WI_patcher = patch('azext_devops.vstsCompressed.work_item_tracking.v4_0.work_item_tracking_client.WorkItemTrackingClient.get_work_item')
         self.create_WI_patcher = patch('azext_devops.vstsCompressed.work_item_tracking.v4_0.work_item_tracking_client.WorkItemTrackingClient.create_work_item')
         self.delete_WI_patcher = patch('azext_devops.vstsCompressed.work_item_tracking.v4_0.work_item_tracking_client.WorkItemTrackingClient.delete_work_item')
-        self.get_credential_patcher = patch('azext_devops.dev.common.services.get_credential')
         self.open_in_browser_patcher = patch('azext_devops.dev.boards.work_item._open_work_item')
-        self.validate_token_patcher = patch('azext_devops.dev.common.services.validate_token_for_instance')
 
         # patch get client so no network call is made
         self.get_client_patcher = patch('azext_devops.vstsCompressed.vss_connection.VssConnection.get_client', new=get_client_mock_helper)
@@ -41,8 +40,6 @@ class TestWorkItemMethods(unittest.TestCase):
         self.mock_get_WI = self.get_WI_patcher.start()
         self.mock_create_WI = self. create_WI_patcher.start()
         self.mock_delete_WI = self.delete_WI_patcher.start()
-        self.mock_get_credential = self.get_credential_patcher.start()
-        self.mock_validate_token = self.validate_token_patcher.start()
         self.mock_open_browser = self.open_in_browser_patcher.start()
 
         #clear connection cache before running each test
@@ -50,12 +47,8 @@ class TestWorkItemMethods(unittest.TestCase):
 
 
     def tearDown(self):
-        self.mock_get_WI.stop()
-        self.mock_create_WI.stop()
-        self.mock_delete_WI.stop()
-        self.mock_get_credential.stop()
-        self.mock_validate_token.stop()
-        self.mock_open_browser.stop()
+        self.authentication_tearDown()
+        patch.stopall()
 
 
     def test_show_work_item_correct_id(self):
@@ -63,8 +56,7 @@ class TestWorkItemMethods(unittest.TestCase):
         test_work_item_id = 1
 
         # set return values
-        self.mock_get_credential.return_value = self._TEST_PAT_TOKEN
-        self.mock_validate_token.return_value = True
+        self.authenticate()
         self.mock_get_WI.return_value.id = test_work_item_id
 
         response = show_work_item(id=test_work_item_id, organization=self._TEST_DEVOPS_ORGANIZATION)
@@ -80,7 +72,7 @@ class TestWorkItemMethods(unittest.TestCase):
         test_work_item_id = 1
 
         # set return values
-        self.mock_validate_token.return_value = True
+        self.authenticate()
         self.mock_get_WI.return_value.id = test_work_item_id
 
         response = show_work_item(id=test_work_item_id, open=True, organization=self._TEST_DEVOPS_ORGANIZATION)
@@ -95,7 +87,7 @@ class TestWorkItemMethods(unittest.TestCase):
 
         test_work_item_id = 1000
 
-        self.mock_validate_token.return_value = True
+        self.authenticate()
         self.mock_get_WI.side_effect = Exception(r'TF401232: Work item 1000 does not exist, or you do not have permissions to read it.')
 
         with self.assertRaises(Exception) as exc:
@@ -112,7 +104,7 @@ class TestWorkItemMethods(unittest.TestCase):
         test_work_item_id = 1
 
         # set return values
-        self.mock_validate_token.return_value = True
+        self.authenticate()
         self.mock_delete_WI.return_value.id = test_work_item_id
 
         response = delete_work_item(id=test_work_item_id, destroy=False, organization=self._TEST_DEVOPS_ORGANIZATION, detect='Off')
@@ -127,7 +119,7 @@ class TestWorkItemMethods(unittest.TestCase):
 
         test_work_item_id = 1000
 
-        self.mock_validate_token.return_value = True
+        self.authenticate()
         self.mock_delete_WI.side_effect = Exception(r'TF401232: Work item 1000 does not exist, or you do not have permissions to read it.')
 
         with self.assertRaises(Exception) as exc:
