+++ azure-devops/azext_devops/test/repos/test_ref.py
@@ -15,24 +15,8 @@ except ImportError:
 from azext_devops.vstsCompressed.git.v4_0.git_client import GitClient
 from azext_devops.dev.common.services import clear_connection_cache
 from azext_devops.dev.repos.ref import (list_refs, create_ref, delete_ref, lock_ref, unlock_ref)
+from azext_devops.test.utils.authentication import AuthenticatedTests
 from azext_devops.test.utils.helper import get_client_mock_helper, TEST_DEVOPS_ORG_URL
-
-
-class AuthenticatedTests(unittest.TestCase):
-
-    def authentication_setUp(self):
-        self.resolve_identity_patcher = patch('azext_devops.dev.common.identities.resolve_identity_as_id')
-        self.get_credential_patcher = patch('azext_devops.dev.common.services.get_credential')
-        self.validate_token_patcher = patch('azext_devops.dev.common.services.validate_token_for_instance')
-
-        # start the patchers
-        self.mock_resolve_identity = self.resolve_identity_patcher.start()
-        self.mock_get_credential = self.get_credential_patcher.start()
-        self.mock_validate_token = self.validate_token_patcher.start()
-
-    def authenticate(self):
-        # set return values
-        self.mock_validate_token.return_value = True
 
 
 class TestRefMethods(AuthenticatedTests):
@@ -56,9 +40,8 @@ class TestRefMethods(AuthenticatedTests):
         clear_connection_cache()
 
     def tearDown(self):
-        self.mock_get_refs.stop()
-        self.mock_update_ref.stop()
-        self.mock_update_refs.stop()
+        self.authentication_tearDown()
+        patch.stopall()
 
     def test_list_refs(self):
         self.authenticate()
