+++ azure-devops/azext_devops/test/repos/test_repository.py
@@ -22,16 +22,17 @@ from azext_devops.dev.repos.repository import (create_repo,
                                                  update_repo)
                                             
 from azext_devops.dev.common.services import clear_connection_cache
+from azext_devops.test.utils.authentication import AuthenticatedTests
+from azext_devops.test.utils.helper import get_client_mock_helper
 
 
-
-class TestRepositoryMethods(unittest.TestCase):
+class TestRepositoryMethods(AuthenticatedTests):
 
     _TEST_DEVOPS_ORGANIZATION = 'https://someorg.visualstudio.com'
     _TEST_PAT_TOKEN = 'lwghjbj67fghokrgxsytghg75nk2ssguljk7a78qpcg2ttygviyt'
 
     def setUp(self):
-
+        self.authentication_setUp()
         self.create_repository_patcher = patch('azext_devops.vstsCompressed.git.v4_0.git_client.GitClient.create_repository')
         self.delete_repository_patcher = patch('azext_devops.vstsCompressed.git.v4_0.git_client.GitClient.delete_repository')
         self.get_repositories_patcher = patch('azext_devops.vstsCompressed.git.v4_0.git_client.GitClient.get_repositories')
@@ -45,16 +46,17 @@ class TestRepositoryMethods(unittest.TestCase):
         self.mock_get_repository = self.get_repository_patcher.start()
         self.mock_update_repository = self.update_repository_patcher.start()
 
+        # Setup mocks for clients
+        self.get_client = patch('azext_devops.vstsCompressed.vss_connection.VssConnection.get_client', new=get_client_mock_helper)
+        self.mock_get_client = self.get_client.start()
+
         #clear connection cache before running each test
         clear_connection_cache()
 
 
     def tearDown(self):
-        self.mock_create_repo.stop()
-        self.mock_delete_repo.stop()
-        self.mock_get_repositories.stop()
-        self.mock_get_repository.stop()
-        self.mock_update_repository.stop()
+        self.authentication_tearDown()
+        patch.stopall()
 
 
     def test_create_repo(self):
